Generic/[Lower|Upper]CaseConstant: minor clean up
The PHP 8.0 namespaced name tokens were already added to the sniff in PR 119 (to make the merge with the old 4.0 branch easier).

This commit now removes the `T_NS_SEPARATOR` and `T_NAMESPACE` tokens which should no longer need to be taken into account what with the change in the namespaced names tokenization in PHPCS 4.0.

The change is already covered by pre-existing tests.