Modernize: Util\Common: use class constant for constant array
As the property was in the public API, the class constant is also public and the property has been deprecated, to be removed in PHPCS 5.0.

Includes changing the array format from a list of strings to an associative array with the key and value for each entry holding the same string to allow for using `isset()` instead of `in_array()`.